Fix int types.

+++ cpp/src/test_main.cpp
@@ -22,7 +22,7 @@ void on_disconnect (ws_t *ws, const char *reason) {
 }
 
 void test_encode_decode() {
-    int data[] = {0, -1, 1, 100, -100, 1000, 10000};
+    int32_t data[] = {0, -1, 1, 100, -100, 1000, 10000};
     char expected[][4] = { {0, 0, 0, 0},
                            {1, 0, 0, 0},
                            {2, 0, 0, 0},
@@ -37,13 +37,13 @@ void test_encode_decode() {
                           {0, 0, 0, 0},
                           {0, 0, 0, 0},
                           {0, 0, 0, 0} };
-    int num_cases = sizeof(data) / sizeof(int);
+    uint_fast8_t num_cases = sizeof(data) / sizeof(int32_t);
 
     cout << "Testing varint-zz encoding:" << endl;
-    for(int i=0; i < num_cases; ++i) {
-        int len = encode_int(data[i], encoded[i]);
+    for(uint_fast8_t i=0; i < num_cases; ++i) {
+        uint_fast8_t len = encode_int(data[i], encoded[i]);
         bool success = true;
-        for(int j=0; j<len; ++j) {
+        for(uint_fast8_t j=0; j<len; ++j) {
             if(expected[i][j] != encoded[i][j]) {
                 success = false;
                 break;
@@ -54,8 +54,8 @@ void test_encode_decode() {
     cout << endl;
 
     cout << "Testing varint-zz decoding:" << endl;
-    for(int i=0; i < num_cases; ++i) {
-        int n = decode_int(encoded[i]);
+    for(uint_fast8_t i=0; i < num_cases; ++i) {
+        int32_t n = decode_int(encoded[i]);
         if(n == data[i]) {
             cout << ".";
         } else {
